gerri/operator/requirements.txt
websockets
pypubsub
--trusted-host 175.126.123.199
--index-url http://keti:dev123@175.126.123.199:7070
ketirtc
pyside6
dynamixel_sdk